day6/leaveapi/requirements.txt
Django>=4.2.0
requests>=2.31.0
djangorestframework>=3.14.0
psycopg2>=2.9.0
django-rest-swagger>=2.2.0
drf-yasg>=1.21.0
setuptools>=65.5.0
hvac>=0.12.0
spring-config-client>=0.1.0
confluent-kafka>=2.3.0
python-dotenv
mongoengine>=0.24.0
django-sslserver